ai/requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.3
joblib==1.4.2
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.27.2
tqdm==4.67.1